[fix]#23 - 프리랜서 마이페이지 경로 수정
